Add trip controller
